Correciónes

+++ Video.cpp
@@ -1,6 +1,6 @@
 #include "Video.h"
 
-Video::Video(string ID, string nombre, string generos, int calificacion, double minutos, string fecha)
+Video::Video(string ID, string nombre, string generos, double calificacion, double minutos, string fecha)
 {
     this->ID = ID;
     this->nombre = nombre;
